kernel for torch gemm
