+++ benchmarker/modules/do_torch.py
@@ -12,6 +12,7 @@ class Benchmark(IGEMM):
     def __init__(self, params, remaining_args=None):
         super().__init__(params, remaining_args)
         self.a, self.b, self.c = self.load_data()
+        self.get_kernel(params, remaining_args)
 
     def run(self):
         if "nb_gpus" in self.params:
