+++ benchmarker/modules/problems/gemm/torch.py
@@ -1,0 +1,12 @@
+import torch
+
+
+class Net:
+    def __call__(data):
+        x, y = data
+        result = x @ y
+        return result
+
+
+def get_kernel(params):
+    return Net
